+++ src/upload-lib.test.ts
@@ -3,9 +3,8 @@ import * as path from "path";
 
 import test from "ava";
 
-import { Feature } from "./feature-flags";
 import { getRunnerLogger, Logger } from "./logging";
-import { createFeatures, setupTests } from "./testing-utils";
+import { setupTests } from "./testing-utils";
 import * as uploadLib from "./upload-lib";
 import { GitHubVariant, initializeEnvironment, withTmpDir } from "./util";
 
@@ -325,23 +324,10 @@ test("accept results with invalid artifactLocation.uri value", (t) => {
   );
 });
 
-test("shouldShowCombineSarifFilesDeprecationWarning when on dotcom with feature flag", async (t) => {
+test("shouldShowCombineSarifFilesDeprecationWarning when on dotcom", async (t) => {
   t.true(
     await uploadLib.shouldShowCombineSarifFilesDeprecationWarning(
       [createMockSarif("abc", "def"), createMockSarif("abc", "def")],
-      createFeatures([Feature.CombineSarifFilesDeprecationWarning]),
-      {
-        type: GitHubVariant.DOTCOM,
-      },
-    ),
-  );
-});
-
-test("shouldShowCombineSarifFilesDeprecationWarning without feature flag", async (t) => {
-  t.false(
-    await uploadLib.shouldShowCombineSarifFilesDeprecationWarning(
-      [createMockSarif("abc", "def"), createMockSarif("abc", "def")],
-      createFeatures([]),
       {
         type: GitHubVariant.DOTCOM,
       },
@@ -353,7 +339,6 @@ test("shouldShowCombineSarifFilesDeprecationWarning when on GHES 3.13", async (t
   t.false(
     await uploadLib.shouldShowCombineSarifFilesDeprecationWarning(
       [createMockSarif("abc", "def"), createMockSarif("abc", "def")],
-      createFeatures([Feature.CombineSarifFilesDeprecationWarning]),
       {
         type: GitHubVariant.GHES,
         version: "3.13.2",
@@ -366,7 +351,6 @@ test("shouldShowCombineSarifFilesDeprecationWarning when on GHES 3.14", async (t
   t.true(
     await uploadLib.shouldShowCombineSarifFilesDeprecationWarning(
       [createMockSarif("abc", "def"), createMockSarif("abc", "def")],
-      createFeatures([Feature.CombineSarifFilesDeprecationWarning]),
       {
         type: GitHubVariant.GHES,
         version: "3.14.0",
@@ -379,7 +363,6 @@ test("shouldShowCombineSarifFilesDeprecationWarning with only 1 run", async (t) 
   t.false(
     await uploadLib.shouldShowCombineSarifFilesDeprecationWarning(
       [createMockSarif("abc", "def")],
-      createFeatures([Feature.CombineSarifFilesDeprecationWarning]),
       {
         type: GitHubVariant.DOTCOM,
       },
@@ -391,7 +374,6 @@ test("shouldShowCombineSarifFilesDeprecationWarning with distinct categories", a
   t.false(
     await uploadLib.shouldShowCombineSarifFilesDeprecationWarning(
       [createMockSarif("abc", "def"), createMockSarif("def", "def")],
-      createFeatures([Feature.CombineSarifFilesDeprecationWarning]),
       {
         type: GitHubVariant.DOTCOM,
       },
@@ -403,7 +385,6 @@ test("shouldShowCombineSarifFilesDeprecationWarning with distinct tools", async 
   t.false(
     await uploadLib.shouldShowCombineSarifFilesDeprecationWarning(
       [createMockSarif("abc", "abc"), createMockSarif("abc", "def")],
-      createFeatures([Feature.CombineSarifFilesDeprecationWarning]),
       {
         type: GitHubVariant.DOTCOM,
       },
@@ -417,7 +398,6 @@ test("shouldShowCombineSarifFilesDeprecationWarning when environment variable is
   t.false(
     await uploadLib.shouldShowCombineSarifFilesDeprecationWarning(
       [createMockSarif("abc", "def"), createMockSarif("abc", "def")],
-      createFeatures([Feature.CombineSarifFilesDeprecationWarning]),
       {
         type: GitHubVariant.DOTCOM,
       },